moved layer data tests to layer data
